+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -361,6 +361,11 @@ New checks
 
   Finds redundant uses of the ``typename`` keyword.
 
+- New :doc:`totto-function-passing-type
+  <clang-tidy/checks/totto/function-passing-type>` check.
+
+  FIXME: Write a short description.
+
 - New :doc:`totto-use-fixed-width-types-var
   <clang-tidy/checks/totto/use-fixed-width-types-var>` check.